androidApp/FirebaseRealtimeDatabase/BackendPy/requirements.txt
PyQt5==5.14.1
PyQt5-sip==12.7.0
opencv-python==4.1.2.30
numpy==1.18.1
imutils==0.5.3
pytesseract==0.3.1
pandas==0.25.3
pyperclip==1.7.0
requests==2.23.0
python-firebase==1.2
